Cleanup before introducing new test
