--- services/web/server/tests/sandbox/openapi.yaml
@@ -1,113 +1,0 @@
-components:
-  schemas:
-    FileType2Viewer:
-      properties:
-        file_type:
-          title: File Type
-          type: string
-        redirection_url:
-          description: Base url to redirect to this viewer. Needs appending file_size,
-            [file_name] and download_link
-          format: uri
-          maxLength: 2083
-          minLength: 1
-          title: Redirection Url
-          type: string
-        viewer_title:
-          description: Short formatted label with name and version of the viewer
-          title: Viewer Title
-          type: string
-      required:
-      - file_type
-      - viewer_title
-      - redirection_url
-      title: FileType2Viewer
-      type: object
-    HTTPValidationError:
-      properties:
-        detail:
-          items:
-            $ref: '#/components/schemas/ValidationError'
-          title: Detail
-          type: array
-      title: HTTPValidationError
-      type: object
-    ValidationError:
-      properties:
-        loc:
-          items:
-            type: string
-          title: Location
-          type: array
-        msg:
-          title: Message
-          type: string
-        type:
-          title: Error Type
-          type: string
-      required:
-      - loc
-      - msg
-      - type
-      title: ValidationError
-      type: object
-info:
-  title: FastAPI
-  version: 0.1.0
-openapi: 3.0.2
-paths:
-  v0/viewers:
-    get:
-      operationId: get_viewer_for_filev0_viewers_get
-      parameters:
-      - in: query
-        name: file_name
-        required: true
-        schema:
-          title: File Name
-          type: string
-      - in: query
-        name: file_size
-        required: true
-        schema:
-          exclusiveMinimum: 0.0
-          title: File Size
-          type: integer
-      - in: query
-        name: file_type
-        required: true
-        schema:
-          title: File Type
-          type: string
-      responses:
-        '200':
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/FileType2Viewer'
-          description: Successful Response
-        '422':
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/HTTPValidationError'
-          description: Validation Error
-      summary: Get Viewer For File
-      tags:
-      - viewer
-  v0/viewers/filetypes:
-    get:
-      operationId: list_supported_filetypesv0_viewers_filetypes_get
-      responses:
-        '200':
-          content:
-            application/json:
-              schema:
-                items:
-                  $ref: '#/components/schemas/FileType2Viewer'
-                title: Response List Supported Filetypesv0 Viewers Filetypes Get
-                type: array
-          description: Successful Response
-      summary: List Supported Filetypes
-      tags:
-      - viewer
